+++ examples/01-cats-app/src/modules/common/pipes/validation.pipe.ts
@@ -1,4 +1,4 @@
-import { HttpException } from '@nestjs/core';
+import { HttpException } from '@nestjs/common';
 import { PipeTransform, Pipe, ArgumentMetadata, HttpStatus } from '@nestjs/common';
 import { validate } from 'class-validator';
 import { plainToClass } from 'class-transformer';