+++ src/core/errors/messages.ts
@@ -1,11 +1,20 @@
+export const UnknownDependenciesMessage = (type: string, index: number, length: number) => {
+  let message = `Nest can't resolve dependencies of the ${type}`;
+  message += ` (`;
+
+  const args = new Array(length).fill('+');
+  args[index] = '?';
+  message += args.join(', ');
+
+  message += `). Please verify whether [${index}] argument is available in the current context.`;
+  return message;
+};
+
 export const InvalidMiddlewareMessage = (name: string) =>
     `The middleware doesn't provide the 'resolve' method (${name})`;
 
 export const InvalidModuleMessage = (scope: string) =>
     `Nest can't create the module instance. The frequent reason of this exception is the circular dependency between modules. Use forwardRef() to avoid it (read more https://docs.nestjs.com/advanced/circular-dependency). Scope [${scope}]`;
-
-export const UnknownDependenciesMessage = (type: string) =>
-    `Nest can't resolve dependencies of the ${type}. Please verify whether all of them are available in the current context.`;
 
 export const UnknownExportMessage = (name: string) =>
     `You are trying to export unknown component (${name}). Remember - your component should be listed both in exports and components arrays!`;